# Initial Apple M1 optimisations