This is roughly a 14% performance improvement on the minify twitter.json benchmark.

These changes are described in more detail at https://github.com/simdjson/simdjson/discussions/1658