@lemire 2021-07-16
That is very clean.