@lemire 2021-07-16
@dougallj Please sync our with main branch and run the following...

```
./build/benchmark/bench_parse_call --benchmark_filter=fast_minify
```

I am getting...

```C++
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        96573 ns        96563 ns           10 Gigabytes=6.54013G/s docs=10.3563k/s
fast_minify_twitter/repeats:10_median      96236 ns        96236 ns           10 Gigabytes=6.56218G/s docs=10.3912k/s
fast_minify_twitter/repeats:10_stddev        538 ns          522 ns           10 Gigabytes=35.2474M/s docs=55.814/s
fast_minify_twitter/repeats:10_max         97419 ns        97329 ns           10 Gigabytes=6.57419G/s docs=10.4102k/s
fast_minify_gsoc/repeats:10_mean          520787 ns       520773 ns           10 Gigabytes=6.39039G/s docs=1.92029k/s
fast_minify_gsoc/repeats:10_median        521033 ns       521025 ns           10 Gigabytes=6.38709G/s docs=1.91929k/s
fast_minify_gsoc/repeats:10_stddev          3176 ns         3164 ns           10 Gigabytes=38.8814M/s docs=11.6837/s
fast_minify_gsoc/repeats:10_max           525216 ns       525217 ns           10 Gigabytes=6.44858G/s docs=1.93777k/s
```

and with your PR...

```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        86799 ns        86798 ns           10 Gigabytes=7.27575G/s docs=11.5211k/s
fast_minify_twitter/repeats:10_median      86735 ns        86735 ns           10 Gigabytes=7.28095G/s docs=11.5293k/s
fast_minify_twitter/repeats:10_stddev        243 ns          244 ns           10 Gigabytes=20.4158M/s docs=32.3284/s
fast_minify_twitter/repeats:10_max         87198 ns        87197 ns           10 Gigabytes=7.31293G/s docs=11.58k/s
fast_minify_gsoc/repeats:10_mean          496170 ns       496155 ns           10 Gigabytes=6.70763G/s docs=2.01562k/s
fast_minify_gsoc/repeats:10_median        496710 ns       496701 ns           10 Gigabytes=6.69986G/s docs=2.01328k/s
fast_minify_gsoc/repeats:10_stddev          4031 ns         4002 ns           10 Gigabytes=53.7471M/s docs=16.1508/s
fast_minify_gsoc/repeats:10_max           505189 ns       505081 ns           10 Gigabytes=6.76932G/s docs=2.03415k/s
```