@lemire 2021-07-16
It seems to cost us a penalty when parsing however...

Current master...

```
❯ ./build/benchmark/bench_parse_call --benchmark_filter=parse_gsoc
Unable to determine clock rate from sysctl: hw.cpufrequency: No such file or directory
2021-07-16 12:00:08
Running ./build/benchmark/bench_parse_call
Run on (8 X 24.121 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB (x8)
  L1 Instruction 128 KiB (x8)
  L2 Unified 4096 KiB (x8)
Load Average: 1.48, 1.71, 1.85
---------------------------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------
parse_gsoc/repeats:10_mean       829218 ns       828689 ns           10 Gigabytes=4.01596G/s docs=1.20678k/s
parse_gsoc/repeats:10_median     829762 ns       829102 ns           10 Gigabytes=4.0138G/s docs=1.20613k/s
parse_gsoc/repeats:10_stddev       6304 ns         5882 ns           10 Gigabytes=28.5008M/s docs=8.56439/s
parse_gsoc/repeats:10_max        839547 ns       837677 ns           10 Gigabytes=4.05294G/s docs=1.21789k/s
```

With this PR...

```
❯ ./build/benchmark/bench_parse_call --benchmark_filter=parse_gsoc
Unable to determine clock rate from sysctl: hw.cpufrequency: No such file or directory
2021-07-16 12:00:17
Running ./build/benchmark/bench_parse_call
Run on (8 X 24.121 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB (x8)
  L1 Instruction 128 KiB (x8)
  L2 Unified 4096 KiB (x8)
Load Average: 1.63, 1.73, 1.86
---------------------------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------
parse_gsoc/repeats:10_mean       891586 ns       891543 ns           10 Gigabytes=3.73279G/s docs=1.12169k/s
parse_gsoc/repeats:10_median     891746 ns       891729 ns           10 Gigabytes=3.7319G/s docs=1.12142k/s
parse_gsoc/repeats:10_stddev       5594 ns         5540 ns           10 Gigabytes=23.195M/s docs=6.97/s
parse_gsoc/repeats:10_max        900747 ns       900499 ns           10 Gigabytes=3.76504G/s docs=1.13138k/s
```