@lemire 2021-07-16
On Demand performance seems unaffected (`./build/benchmark/bench_ondemand --benchmark_filter=ondemand`) or the effect is small.