@lemire 2021-07-16
Right so the regression in parsing performance seems to be due to the introduction of the additional branch.