@lemire 2021-07-16
If I only keep "Avoid some expensive popcounts on ARM", which leaves the parsing performance unchanged (since it only touches the minifier), then I no longer find any benefit to this PR...

Current master...

```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        96573 ns        96563 ns           10 Gigabytes=6.54013G/s docs=10.3563k/s
fast_minify_twitter/repeats:10_median      96236 ns        96236 ns           10 Gigabytes=6.56218G/s docs=10.3912k/s
fast_minify_twitter/repeats:10_stddev        538 ns          522 ns           10 Gigabytes=35.2474M/s docs=55.814/s
fast_minify_twitter/repeats:10_max         97419 ns        97329 ns           10 Gigabytes=6.57419G/s docs=10.4102k/s
fast_minify_gsoc/repeats:10_mean          520787 ns       520773 ns           10 Gigabytes=6.39039G/s docs=1.92029k/s
fast_minify_gsoc/repeats:10_median        521033 ns       521025 ns           10 Gigabytes=6.38709G/s docs=1.91929k/s
fast_minify_gsoc/repeats:10_stddev          3176 ns         3164 ns           10 Gigabytes=38.8814M/s docs=11.6837/s
fast_minify_gsoc/repeats:10_max           525216 ns       525217 ns           10 Gigabytes=6.44858G/s docs=1.93777k/s
```

This PR...

```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        96486 ns        96475 ns           10 Gigabytes=6.54606G/s docs=10.3656k/s
fast_minify_twitter/repeats:10_median      96387 ns        96365 ns           10 Gigabytes=6.5534G/s docs=10.3773k/s
fast_minify_twitter/repeats:10_stddev        478 ns          469 ns           10 Gigabytes=31.7527M/s docs=50.2802/s
fast_minify_twitter/repeats:10_max         97166 ns        97144 ns           10 Gigabytes=6.57985G/s docs=10.4192k/s
fast_minify_gsoc/repeats:10_mean          520446 ns       520112 ns           10 Gigabytes=6.39868G/s docs=1.92278k/s
fast_minify_gsoc/repeats:10_median        521538 ns       521383 ns           10 Gigabytes=6.38269G/s docs=1.91797k/s
fast_minify_gsoc/repeats:10_stddev          4491 ns         4242 ns           10 Gigabytes=52.4204M/s docs=15.7521/s
fast_minify_gsoc/repeats:10_max           525791 ns       524863 ns           10 Gigabytes=6.50058G/s docs=1.9534k/s
```