@lemire 2021-07-16
@dougallj Can you run my tests and see whether you get the same results?


When comparing this PR with our main branch using GCC 10, the result is more mixed (a gain on twitter and a slim regression over gsoc)....

Main branch (GCC 10):

```
$ ./buildgcc/benchmark/bench_parse_call --benchmark_filter=fast_minify
2021-07-16 16:28:30
Running ./buildgcc/benchmark/bench_parse_call
Run on (4 X 48 MHz CPU s)
Load Average: 0.25, 0.15, 0.06
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        97196 ns        97170 ns           10 Gigabytes=6.4993G/s docs=10.2916k/s
fast_minify_twitter/repeats:10_median      97363 ns        97269 ns           10 Gigabytes=6.49248G/s docs=10.2808k/s
fast_minify_twitter/repeats:10_stddev        612 ns          605 ns           10 Gigabytes=40.5411M/s docs=64.1965/s
fast_minify_twitter/repeats:10_max         98088 ns        98088 ns           10 Gigabytes=6.56717G/s docs=10.3991k/s
fast_minify_gsoc/repeats:10_mean          520058 ns       520031 ns           10 Gigabytes=6.39952G/s docs=1.92303k/s
fast_minify_gsoc/repeats:10_median        520295 ns       520258 ns           10 Gigabytes=6.3965G/s docs=1.92212k/s
fast_minify_gsoc/repeats:10_stddev          3283 ns         3282 ns           10 Gigabytes=40.4293M/s docs=12.1488/s
fast_minify_gsoc/repeats:10_max           524633 ns       524635 ns           10 Gigabytes=6.46494G/s docs=1.94269k/s
```

This PR (GCC10):

```
$ ./buildgcc/benchmark/bench_parse_call --benchmark_filter=fast_minify
2021-07-16 16:40:34
Running ./buildgcc/benchmark/bench_parse_call
Run on (4 X 48 MHz CPU s)
Load Average: 0.02, 0.17, 0.11
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        89628 ns        89604 ns           10 Gigabytes=7.04808G/s docs=11.1606k/s
fast_minify_twitter/repeats:10_median      89499 ns        89446 ns           10 Gigabytes=7.06031G/s docs=11.18k/s
fast_minify_twitter/repeats:10_stddev        514 ns          524 ns           10 Gigabytes=41.1306M/s docs=65.13/s
fast_minify_twitter/repeats:10_max         90360 ns        90350 ns           10 Gigabytes=7.10762G/s docs=11.2549k/s
fast_minify_gsoc/repeats:10_mean          524215 ns       524204 ns           10 Gigabytes=6.34865G/s docs=1.90774k/s
fast_minify_gsoc/repeats:10_median        524734 ns       524721 ns           10 Gigabytes=6.34211G/s docs=1.90578k/s
fast_minify_gsoc/repeats:10_stddev          3772 ns         3763 ns           10 Gigabytes=45.6221M/s docs=13.7093/s
fast_minify_gsoc/repeats:10_max           529781 ns       529725 ns           10 Gigabytes=6.41321G/s docs=1.92714k/s
```


If I remove the extra branch and keep only the population count change, then I get a net regression...

```
$ ./buildgcc/benchmark/bench_parse_call --benchmark_filter=fast_minify
2021-07-16 16:43:46
Running ./buildgcc/benchmark/bench_parse_call
Run on (4 X 48 MHz CPU s)
Load Average: 0.17, 0.14, 0.10
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        99033 ns        99032 ns           10 Gigabytes=6.37725G/s docs=10.0983k/s
fast_minify_twitter/repeats:10_median      99023 ns        99022 ns           10 Gigabytes=6.37753G/s docs=10.0988k/s
fast_minify_twitter/repeats:10_stddev        763 ns          763 ns           10 Gigabytes=49.126M/s docs=77.7907/s
fast_minify_twitter/repeats:10_max        100106 ns       100102 ns           10 Gigabytes=6.45693G/s docs=10.2245k/s
fast_minify_gsoc/repeats:10_mean          538648 ns       538630 ns           10 Gigabytes=6.17861G/s docs=1.85665k/s
fast_minify_gsoc/repeats:10_median        538970 ns       538958 ns           10 Gigabytes=6.17457G/s docs=1.85543k/s
fast_minify_gsoc/repeats:10_stddev          3856 ns         3849 ns           10 Gigabytes=44.1158M/s docs=13.2566/s
fast_minify_gsoc/repeats:10_max           544002 ns       543977 ns           10 Gigabytes=6.23228G/s docs=1.87278k/s
```