@dougallj 2021-07-16
Main branch (clang):

```
---------------------------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------
parse_gsoc/repeats:10_mean       814927 ns       814927 ns           10 Gigabytes=4.08359G/s docs=1.2271k/s
parse_gsoc/repeats:10_median     814942 ns       814942 ns           10 Gigabytes=4.08352G/s docs=1.22708k/s
parse_gsoc/repeats:10_stddev        345 ns          345 ns           10 Gigabytes=1.72783M/s docs=0.519206/s
parse_gsoc/repeats:10_max        815450 ns       815450 ns           10 Gigabytes=4.08631G/s docs=1.22792k/s
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        95852 ns        95852 ns           10 Gigabytes=6.58842G/s docs=10.4327k/s
fast_minify_twitter/repeats:10_median      95863 ns        95863 ns           10 Gigabytes=6.58766G/s docs=10.4315k/s
fast_minify_twitter/repeats:10_stddev       53.0 ns         53.1 ns           10 Gigabytes=3.64693M/s docs=5.7749/s
fast_minify_twitter/repeats:10_max         95926 ns        95927 ns           10 Gigabytes=6.59366G/s docs=10.441k/s
fast_minify_gsoc/repeats:10_mean          514582 ns       514582 ns           10 Gigabytes=6.46705G/s docs=1.94332k/s
fast_minify_gsoc/repeats:10_median        514612 ns       514612 ns           10 Gigabytes=6.46669G/s docs=1.94321k/s
fast_minify_gsoc/repeats:10_stddev          90.5 ns         90.5 ns           10 Gigabytes=1.13804M/s docs=0.341975/s
fast_minify_gsoc/repeats:10_max           514693 ns       514693 ns           10 Gigabytes=6.46884G/s docs=1.94386k/s
```

This PR (clang):

```
---------------------------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------
parse_gsoc/repeats:10_mean       885063 ns       885063 ns           10 Gigabytes=3.76G/s docs=1.12986k/s
parse_gsoc/repeats:10_median     885081 ns       885082 ns           10 Gigabytes=3.75991G/s docs=1.12984k/s
parse_gsoc/repeats:10_stddev        693 ns          693 ns           10 Gigabytes=2.94363M/s docs=0.884548/s
parse_gsoc/repeats:10_max        886335 ns       886335 ns           10 Gigabytes=3.76413G/s docs=1.13111k/s
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        84833 ns        84833 ns           10 Gigabytes=7.44423G/s docs=11.7879k/s
fast_minify_twitter/repeats:10_median      84827 ns        84827 ns           10 Gigabytes=7.44473G/s docs=11.7887k/s
fast_minify_twitter/repeats:10_stddev       40.3 ns         40.3 ns           10 Gigabytes=3.53931M/s docs=5.60448/s
fast_minify_twitter/repeats:10_max         84910 ns        84910 ns           10 Gigabytes=7.44806G/s docs=11.794k/s
fast_minify_gsoc/repeats:10_mean          504146 ns       504146 ns           10 Gigabytes=6.60097G/s docs=1.98357k/s
fast_minify_gsoc/repeats:10_median        503434 ns       503433 ns           10 Gigabytes=6.61027G/s docs=1.98636k/s
fast_minify_gsoc/repeats:10_stddev          1336 ns         1336 ns           10 Gigabytes=17.4779M/s docs=5.25204/s
fast_minify_gsoc/repeats:10_max           505830 ns       505829 ns           10 Gigabytes=6.62081G/s docs=1.98953k/s
```

I'll set up docker sometime in the next few days - that sounds really useful, thanks!